requirements.txt
anyio==3.7.1
asgiref==3.7.1
certifi==2023.5.7
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.1.0
cryptography==41.0.7
distlib==0.3.7
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.4
flake8==6.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.5.29
idna==3.4
iniconfig==2.0.0
Markdown==3.4.3
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.0.0
pre-commit==3.4.0
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.7.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.3.2
pytest-django==4.5.2
pytest-mock==3.11.1
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.3
virtualenv==20.24.5
